+++ packages/lit-html/src/test/directives/keyed_test.ts
@@ -7,7 +7,7 @@
 import {keyed} from 'lit-html/directives/keyed.js';
 import {html, render} from 'lit-html';
 import {stripExpressionMarkers} from '@lit-labs/testing';
-import {assert} from '@esm-bundle/chai';
+import {assert} from 'chai';
 
 suite('keyed directive', () => {
   let container: HTMLDivElement;