+++ packages/reactive-element/src/test/reactive-element_test.ts
@@ -13,7 +13,7 @@ import {
   ReactiveElement,
 } from '@lit/reactive-element';
 import {generateElementName, nextFrame} from './test-helpers.js';
-import {assert} from '@esm-bundle/chai';
+import {assert} from 'chai';
 
 // Note, since tests are not built with production support, detect DEV_MODE
 // by checking if warning API is available.