+++ packages/context/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @lit-labs/context
 
+## 1.1.2
+
+### Patch Changes
+
+- [#4598](https://github.com/lit/lit/pull/4598) [`7cfb2274`](https://github.com/lit/lit/commit/7cfb2274d8903c5be40ca4dcaf5167a57bb36a32) Thanks [@jun-sheaf](https://github.com/jun-sheaf)! - Use `target.set` instead of member assignment in `@consume()` decorator.
+
+## 1.1.1
+
+### Patch Changes
+
+- [#4614](https://github.com/lit/lit/pull/4614) [`ab77cc9c`](https://github.com/lit/lit/commit/ab77cc9c83a67b60b42e77cf736fefbc6503f824) - Fix context types to be compatible with WCCG context protocol types.
+
 ## 1.1.0
 
 ### Minor Changes